Removing a couple of integration test classes having only one test, and easy to merge in another
